Split cast::dictionary into a submodule of cast
